more progress on resolving circular import issues
